requirements.txt
google-genai
dotenv
pandas
fsspec
huggingface_hub